image-gen/requirements.txt
torch==2.5.*
--extra-index-url https://download.pytorch.org/whl/cpu
diffusers==0.30.3
transformers==4.44.2
accelerate==0.33.0
safetensors
sentencepiece
protobuf
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
